icd -> initial_contacts
