+++ docs/conf.py
@@ -201,7 +201,7 @@ sphinx_gallery_conf = {
             "../examples/data",
             "../examples/pipeline",
             "../examples/gsd",
-            "../examples/icd",
+            "../examples/initial_contacts",
             "../examples/lrc",
             "../examples/cadence",
             "../examples/stride_length",
