+++ examples/initial_contacts/README.rst
