+++ examples/initial_contacts/_01_icd_ionescu.py
@@ -1,29 +1,37 @@
-"""
-HKLee algo
-==========
+r"""
+.. _icd_ionescu:
 
-This example shows how to use the improved HKLee algorithm and some examples on how the results compare to the original
+ICD Ionescu
+===========
+
+This example shows how to use the initial_contacts Ionescu algorithm and how its results compare to the original
 matlab implementation.
 
 """
 
+# %%
+# Import useful modules and packages
+import matplotlib.pyplot as plt
 import pandas as pd
-from matplotlib import pyplot as plt
 from mobgap.data import LabExampleDataset
-from mobgap.icd._hklee_algo_improved import IcdHKLeeImproved
+from mobgap.initial_contacts import IcdIonescu
 
 # %%
-# Loading data
-# ------------
+# Loading some example data
+# -------------------------
 # .. note :: More infos about data loading can be found in the :ref:`data loading example <data_loading_example>`.
 #
 # We load example data from the lab dataset together with the INDIP reference system.
-# We will use the INDIP output for initial contacts ("ic") as ground truth.
+# We will use the INDIP "InitialContact_Event" output as ground truth.
 
 example_data = LabExampleDataset(
     reference_system="INDIP", reference_para_level="wb"
-)
+)  # alternatively: "StereoPhoto"
 
+# %%
+# Performance on a single lab trial
+# ---------------------------------
+# Below we apply the algorithm to a lab trail, where we only expect a single gait sequence.
 single_test = example_data.get_subset(
     cohort="HA", participant_id="001", test="Test11", trial="Trial1"
 )
@@ -34,6 +42,7 @@ sampling_rate_hz = single_test.sampling_rate_hz
 ref_ics = single_test.reference_parameters_.ic_list
 
 reference_wbs
+
 # %%
 # Applying the algorithm
 # ----------------------
@@ -45,14 +54,12 @@ iterator = GsIterator()
 
 for (gs, data), result in iterator.iterate(imu_data, reference_wbs):
     result.ic_list = (
-        IcdHKLeeImproved()
-        .detect(data, sampling_rate_hz=sampling_rate_hz)
-        .ic_list_
+        IcdIonescu().detect(data, sampling_rate_hz=sampling_rate_hz).ic_list_
     )
 
 detected_ics = iterator.results_.ic_list
-
 detected_ics
+
 # %%
 # Matlab Outputs
 # --------------
@@ -66,7 +73,7 @@ def load_matlab_output(datapoint):
     p = datapoint.group_label
     with (
         PACKAGE_ROOT.parent
-        / f"example_data/original_results/icd_hklee_improved/lab/{p.cohort}/{p.participant_id}/SD_Output.json"
+        / f"example_data/original_results/icd_ionescu/lab/{p.cohort}/{p.participant_id}/SD_Output.json"
     ).open() as f:
         original_results = json.load(f)["SD_Output"][p.time_measure][p.test][
             p.trial
@@ -92,20 +99,15 @@ detected_ics_matlab
 # --------------------
 # With that we can compare the python, matlab and ground truth results.
 # We zoom in into one of the gait sequences to better see the output.
-#
 # We can make a couple of main observations:
 #
-# 1. The python version finds (for the most part) the same ICs as the matlab version, but detects them slightly earlies
-#    (around 2-8 samples/20-80 ms).
+# 1. The python version finds the same ICs as the matlab version, but wil a small shift to the left (around 2-5
+#    samples/20-50 ms).
 #    This is likely due to some differences in the downsampling process.
-# 2. Matlab and Python sometimes differ in the detection in the first and the last IC, if they are close to the border
-#    of the GS.
-#    This is because the morphological filter in the Python and the Matlab treat the border differently.
-# 3. Compared to the ground truth reference, both versions detect the IC too late most of the time.
-# 4. Both algorithms can not detect the first IC of the gait sequence.
+# 2. Compared to the ground truth reference, both versions detect the IC too early most of the time.
+# 3. Both algorithms can not detect the first IC of the gait sequence.
 #    However, this is expected, as per definition, this first IC marks the start of the WB in the reference system.
 #    Hence, there are no samples before that point the algorithm can use to detect the IC.
-
 imu_data.reset_index(drop=True).plot(y="acc_x")
 
 plt.plot(ref_ics["ic"], imu_data["acc_x"].iloc[ref_ics["ic"]], "o", label="ref")
@@ -113,15 +115,15 @@ plt.plot(
     detected_ics["ic"],
     imu_data["acc_x"].iloc[detected_ics["ic"]],
     "x",
-    label="hklee_algo_py",
+    label="icd_ionescu_py",
 )
 plt.plot(
     detected_ics_matlab["ic"],
     imu_data["acc_x"].iloc[detected_ics_matlab["ic"]],
     "+",
-    label="hklee_algo_matlab",
+    label="icd_ionescu_matlab",
 )
-plt.xlim(reference_wbs.iloc[3]["start"] - 50, reference_wbs.iloc[3]["end"] + 50)
+plt.xlim(reference_wbs.iloc[2]["start"] - 50, reference_wbs.iloc[2]["end"] + 50)
 plt.legend()
 plt.show()
