+++ examples/initial_contacts/_03_hklee_algo.py
@@ -1,37 +1,29 @@
-r"""
-.. _icd_ionescu:
+"""
+HKLee algo
+==========
 
-ICD Ionescu
-===========
-
-This example shows how to use the icd Ionescu algorithm and how its results compare to the original
+This example shows how to use the improved HKLee algorithm and some examples on how the results compare to the original
 matlab implementation.
 
 """
 
-# %%
-# Import useful modules and packages
-import matplotlib.pyplot as plt
 import pandas as pd
+from matplotlib import pyplot as plt
 from mobgap.data import LabExampleDataset
-from mobgap.icd import IcdIonescu
+from mobgap.initial_contacts._hklee_algo_improved import IcdHKLeeImproved
 
 # %%
-# Loading some example data
-# -------------------------
+# Loading data
+# ------------
 # .. note :: More infos about data loading can be found in the :ref:`data loading example <data_loading_example>`.
 #
 # We load example data from the lab dataset together with the INDIP reference system.
-# We will use the INDIP "InitialContact_Event" output as ground truth.
+# We will use the INDIP output for initial contacts ("ic") as ground truth.
 
 example_data = LabExampleDataset(
     reference_system="INDIP", reference_para_level="wb"
-)  # alternatively: "StereoPhoto"
+)
 
-# %%
-# Performance on a single lab trial
-# ---------------------------------
-# Below we apply the algorithm to a lab trail, where we only expect a single gait sequence.
 single_test = example_data.get_subset(
     cohort="HA", participant_id="001", test="Test11", trial="Trial1"
 )
@@ -42,7 +34,6 @@ sampling_rate_hz = single_test.sampling_rate_hz
 ref_ics = single_test.reference_parameters_.ic_list
 
 reference_wbs
-
 # %%
 # Applying the algorithm
 # ----------------------
@@ -54,12 +45,14 @@ iterator = GsIterator()
 
 for (gs, data), result in iterator.iterate(imu_data, reference_wbs):
     result.ic_list = (
-        IcdIonescu().detect(data, sampling_rate_hz=sampling_rate_hz).ic_list_
+        IcdHKLeeImproved()
+        .detect(data, sampling_rate_hz=sampling_rate_hz)
+        .ic_list_
     )
 
 detected_ics = iterator.results_.ic_list
-detected_ics
 
+detected_ics
 # %%
 # Matlab Outputs
 # --------------
@@ -73,7 +66,7 @@ def load_matlab_output(datapoint):
     p = datapoint.group_label
     with (
         PACKAGE_ROOT.parent
-        / f"example_data/original_results/icd_ionescu/lab/{p.cohort}/{p.participant_id}/SD_Output.json"
+        / f"example_data/original_results/icd_hklee_improved/lab/{p.cohort}/{p.participant_id}/SD_Output.json"
     ).open() as f:
         original_results = json.load(f)["SD_Output"][p.time_measure][p.test][
             p.trial
@@ -99,15 +92,20 @@ detected_ics_matlab
 # --------------------
 # With that we can compare the python, matlab and ground truth results.
 # We zoom in into one of the gait sequences to better see the output.
+#
 # We can make a couple of main observations:
 #
-# 1. The python version finds the same ICs as the matlab version, but wil a small shift to the left (around 2-5
-#    samples/20-50 ms).
+# 1. The python version finds (for the most part) the same ICs as the matlab version, but detects them slightly earlies
+#    (around 2-8 samples/20-80 ms).
 #    This is likely due to some differences in the downsampling process.
-# 2. Compared to the ground truth reference, both versions detect the IC too early most of the time.
-# 3. Both algorithms can not detect the first IC of the gait sequence.
+# 2. Matlab and Python sometimes differ in the detection in the first and the last IC, if they are close to the border
+#    of the GS.
+#    This is because the morphological filter in the Python and the Matlab treat the border differently.
+# 3. Compared to the ground truth reference, both versions detect the IC too late most of the time.
+# 4. Both algorithms can not detect the first IC of the gait sequence.
 #    However, this is expected, as per definition, this first IC marks the start of the WB in the reference system.
 #    Hence, there are no samples before that point the algorithm can use to detect the IC.
+
 imu_data.reset_index(drop=True).plot(y="acc_x")
 
 plt.plot(ref_ics["ic"], imu_data["acc_x"].iloc[ref_ics["ic"]], "o", label="ref")
@@ -115,15 +113,15 @@ plt.plot(
     detected_ics["ic"],
     imu_data["acc_x"].iloc[detected_ics["ic"]],
     "x",
-    label="icd_ionescu_py",
+    label="hklee_algo_py",
 )
 plt.plot(
     detected_ics_matlab["ic"],
     imu_data["acc_x"].iloc[detected_ics_matlab["ic"]],
     "+",
-    label="icd_ionescu_matlab",
+    label="hklee_algo_matlab",
 )
-plt.xlim(reference_wbs.iloc[2]["start"] - 50, reference_wbs.iloc[2]["end"] + 50)
+plt.xlim(reference_wbs.iloc[3]["start"] - 50, reference_wbs.iloc[3]["end"] + 50)
 plt.legend()
 plt.show()
