+++ examples/initial_contacts/_04_icd_evaluation.py
@@ -12,7 +12,7 @@ import pandas as pd
 # %%
 # Import useful modules and packages
 from mobgap.data import LabExampleDataset
-from mobgap.icd import IcdIonescu
+from mobgap.initial_contacts import IcdIonescu
 from mobgap.pipeline import GsIterator
 
 # %%
@@ -78,7 +78,7 @@ reference_ics
 # Let's quantify how the algorithm output compares to the reference labels.
 # To gain a detailed insight into the performance of the algorithm, we can look into the individual matches between the
 # detected and reference initial contacts.
-# To do this, we use the :func:`~mobgap.icd.evaluation.categorize_ic_list` function to classify each detected initial
+# To do this, we use the :func:`~mobgap.initial_contacts.evaluation.categorize_ic_list` function to classify each detected initial
 # contact as a true positive, false positive, or false negative.
 # We can then use these results to calculate a range of higher-level performance metrics.
 #
@@ -103,7 +103,7 @@ per_wb_grouper = create_multi_groupby(
 # I.e. the function will get the detected and reference initial contacts for each walking bout and then can perform
 # some operation on them.
 #
-# In our case we want to apply the :func:`~mobgap.icd.evaluation.categorize_ic_list` function to each walking bout.
+# In our case we want to apply the :func:`~mobgap.initial_contacts.evaluation.categorize_ic_list` function to each walking bout.
 # This function will then return a dataframe with the matches given a certain tolerance.
 #
 # We don't assume that initial contacts are detected at perfectly the exact same time in both systems.
@@ -121,7 +121,7 @@ tolerance_samples
 # and the match type.
 # The two index columns contain tuples in our case, as they stem from the original multiindex that we provided.
 # So each of the tuples has the form ``(wb_id, ic_id)``.
-from mobgap.icd.evaluation import categorize_ic_list
+from mobgap.initial_contacts.evaluation import categorize_ic_list
 
 matches_per_wb = create_multi_groupby(
     detected_ics, reference_ics, groupby="wb_id"
@@ -163,7 +163,7 @@ matched_all
 # -------------------------------
 # From these ``matches_per_wb``, a range of higher-level performance metrics (including the total number of true
 # positives, false positives, and false negatives, as well as precision, recall, and F1-score) can be calculated.
-# For this purpose, we can use the :func:`~mobgap.icd.evaluation.calculate_matched_icd_performance_metrics` function.
+# For this purpose, we can use the :func:`~mobgap.initial_contacts.evaluation.calculate_matched_icd_performance_metrics` function.
 # It returns a dictionary containing all metrics for the specified detected and reference initial contact lists.
 #
 # We can again decide, if we want to calculate these metrics across all walking bouts or for each walking bout
@@ -171,7 +171,9 @@ matched_all
 # We will quickly show both approaches below.
 #
 # Across all walking bouts:
-from mobgap.icd.evaluation import calculate_matched_icd_performance_metrics
+from mobgap.initial_contacts.evaluation import (
+    calculate_matched_icd_performance_metrics,
+)
 
 metrics_all = calculate_matched_icd_performance_metrics(matches_per_wb)
