--- mobgap/icd/__init__.py
@@ -1,8 +1,0 @@
-"""Algorithms to detect ICs within raw IMU data during a gait sequence."""
-
-from mobgap.icd._hklee_algo_improved import IcdHKLeeImproved
-from mobgap.icd._icd_ionescu import IcdIonescu
-from mobgap.icd._shin_algo_improved import IcdShinImproved
-from mobgap.icd._utils import refine_gs
-
-__all__ = ["IcdShinImproved", "IcdIonescu", "IcdHKLeeImproved", "refine_gs"]
