+++ mobgap/initial_contacts/__init__.py
@@ -1,0 +1,8 @@
+"""Algorithms to detect ICs within raw IMU data during a gait sequence."""
+
+from mobgap.initial_contacts._hklee_algo_improved import IcdHKLeeImproved
+from mobgap.initial_contacts._icd_ionescu import IcdIonescu
+from mobgap.initial_contacts._shin_algo_improved import IcdShinImproved
+from mobgap.initial_contacts._utils import refine_gs
+
+__all__ = ["IcdShinImproved", "IcdIonescu", "IcdHKLeeImproved", "refine_gs"]
