+++ mobgap/initial_contacts/base.py
