+++ mobgap/utils/evaluation.py
@@ -41,7 +41,7 @@ def precision_recall_f1_score(
         Currently supported are dataframes resulting either from the evaluation of gait sequence detection algorithms
         (results from :func:`~mobgap.gsd.evaluation.categorize_intervals_per_sample`),
         or from the evaluation of initial contact detection algorithms
-        (results from :func:`~mobgap.icd.evaluation.categorize_ic_list`).
+        (results from :func:`~mobgap.initial_contacts.evaluation.categorize_ic_list`).
         To be handled like a gait sequence detection evaluation dataframe,
         the `matches_df` must have columns "start", "end", and "match_type".
         To be handled like an initial contact detection evaluation dataframe,
@@ -86,7 +86,7 @@ def precision_score(matches_df: pd.DataFrame, *, zero_division: Literal["warn", 
         Currently supported are dataframes resulting either from the evaluation of gait sequence detection algorithms
         (results from :func:`~mobgap.gsd.evaluation.categorize_intervals_per_sample`),
         or from the evaluation of initial contact detection algorithms
-        (results from :func:`~mobgap.icd.evaluation.categorize_ic_list`).
+        (results from :func:`~mobgap.initial_contacts.evaluation.categorize_ic_list`).
         To be handled like a gait sequence detection evaluation dataframe,
         the `matches_df` must have columns "start", "end", and "match_type".
         To be handled like an initial contact detection evaluation dataframe,
@@ -140,7 +140,7 @@ def recall_score(matches_df: pd.DataFrame, *, zero_division: Literal["warn", 0, 
         Currently supported are dataframes resulting either from the evaluation of gait sequence detection algorithms
         (results from :func:`~mobgap.gsd.evaluation.categorize_intervals_per_sample`),
         or from the evaluation of initial contact detection algorithms
-        (results from :func:`~mobgap.icd.evaluation.categorize_ic_list`).
+        (results from :func:`~mobgap.initial_contacts.evaluation.categorize_ic_list`).
         To be handled like a gait sequence detection evaluation dataframe,
         the `matches_df` must have columns "start", "end", and "match_type".
         To be handled like an initial contact detection evaluation dataframe,
@@ -399,7 +399,7 @@ def f1_score(matches_df: pd.DataFrame, *, zero_division: Literal["warn", 0, 1] =
         Currently supported are dataframes resulting either from the evaluation of gait sequence detection algorithms
         (results from :func:`~mobgap.gsd.evaluation.categorize_intervals_per_sample`),
         or from the evaluation of initial contact detection algorithms
-        (results from :func:`~mobgap.icd.evaluation.categorize_ic_list`).
+        (results from :func:`~mobgap.initial_contacts.evaluation.categorize_ic_list`).
         To be handled like a gait sequence detection evaluation dataframe,
         the `matches_df` must have columns "start", "end", and "match_type".
         To be handled like an initial contact detection evaluation dataframe,
