+++ tests/test_examples/test_initial_contacts.py
@@ -6,25 +6,25 @@ matplotlib.use("Agg")
 
 
 def test_icd_ionescu(snapshot):
-    from examples.icd._01_icd_ionescu import detected_ics
+    from examples.initial_contacts._01_icd_ionescu import detected_ics
 
     snapshot.assert_match(detected_ics, "short_trial_output")
 
 
 def test_icd_shin_improved(snapshot):
-    from examples.icd._02_shin_algo import detected_ics
+    from examples.initial_contacts._02_shin_algo import detected_ics
 
     snapshot.assert_match(detected_ics, "short_trial_output")
 
 
 def test_icd_hklee_improved(snapshot):
-    from examples.icd._03_hklee_algo import detected_ics
+    from examples.initial_contacts._03_hklee_algo import detected_ics
 
     snapshot.assert_match(detected_ics, "short_trial_output")
 
 
 def test_icd_evaluation(snapshot):
-    from examples.icd._04_icd_evaluation import matches_per_wb, metrics_all
+    from examples.initial_contacts._04_icd_evaluation import matches_per_wb, metrics_all
 
     snapshot.assert_match(matches_per_wb, "matches")
     snapshot.assert_match(pd.DataFrame(metrics_all, index=[0]), "metrics")
