+++ tests/test_initial_contacts/snapshot/TestCalculateIcdMetrics.test_ic_list_input_metrics.json
