+++ tests/test_initial_contacts/snapshot/TestCalculateIcdMetrics.test_multiindex_ic_list_input_metrics.json
