+++ tests/test_initial_contacts/snapshot/TestHKLeeImprovedRegression.test_example_lab_data[datapoint1]_('HA', '001', 'TimeMeasure1', 'Test5', 'Trial2').json
