+++ tests/test_initial_contacts/snapshot/TestIcdIonescuRegression.test_example_lab_data[datapoint2]_('HA', '001', 'TimeMeasure1', 'Test11', 'Trial1').json
