+++ tests/test_initial_contacts/snapshot/TestIcdIonescuRegression.test_example_lab_data[datapoint5]_('HA', '002', 'TimeMeasure1', 'Test11', 'Trial1').json
