+++ tests/test_initial_contacts/snapshot/TestIcdIonescuRegression.test_example_lab_data[datapoint8]_('MS', '001', 'TimeMeasure1', 'Test11', 'Trial1').json
