+++ tests/test_initial_contacts/snapshot/TestShinImprovedRegression.test_example_lab_data[datapoint6]_('MS', '001', 'TimeMeasure1', 'Test5', 'Trial1').json
