+++ tests/test_initial_contacts/test_icd_hklee_improved.py
@@ -4,37 +4,37 @@ import pytest
 from tpcp.testing import TestAlgorithmMixin
 
 from mobgap.data import LabExampleDataset
-from mobgap.icd import IcdShinImproved
+from mobgap.initial_contacts._hklee_algo_improved import IcdHKLeeImproved
 from mobgap.pipeline import GsIterator
 
 
-class TestMetaShinImproved(TestAlgorithmMixin):
+class TestMetaHKLeeImproved(TestAlgorithmMixin):
     __test__ = True
 
-    ALGORITHM_CLASS = IcdShinImproved
+    ALGORITHM_CLASS = IcdHKLeeImproved
 
     @pytest.fixture()
     def after_action_instance(self):
         return self.ALGORITHM_CLASS().detect(
-            pd.DataFrame(np.zeros((1000, 3)), columns=["acc_x", "acc_y", "acc_z"]), sampling_rate_hz=40.0
+            pd.DataFrame(np.zeros((1000, 3)), columns=["acc_x", "acc_y", "acc_z"]), sampling_rate_hz=120.0
         )
 
 
-class TestShinImproved:
+class TestHKLeeImproved:
     def test_invalid_axis_parameter(self):
         with pytest.raises(ValueError):
-            IcdShinImproved(axis="invalid").detect(pd.DataFrame(), sampling_rate_hz=100)
+            IcdHKLeeImproved(axis="invalid").detect(pd.DataFrame(), sampling_rate_hz=100)
 
     def test_no_ics_detected(self):
         data = pd.DataFrame(np.zeros((1000, 3)), columns=["acc_x", "acc_y", "acc_z"])
-        output = IcdShinImproved(axis="x")
-        output.detect(data, sampling_rate_hz=40.0)
+        output = IcdHKLeeImproved(axis="x")
+        output.detect(data, sampling_rate_hz=120.0)
         output_ic = output.ic_list_["ic"]
         empty_output = {}
         assert output_ic.to_dict() == empty_output
 
 
-class TestShinImprovedRegression:
+class TestHKLeeImprovedRegression:
     @pytest.mark.parametrize("datapoint", LabExampleDataset(reference_system="INDIP", reference_para_level="wb"))
     def test_example_lab_data(self, datapoint, snapshot):
         data = datapoint.data_ss
@@ -46,7 +46,7 @@ class TestShinImprovedRegression:
         iterator = GsIterator()
 
         for (gs, data), result in iterator.iterate(data, ref_walk_bouts):
-            result.ic_list = IcdShinImproved().detect(data, sampling_rate_hz=sampling_rate_hz).ic_list_
+            result.ic_list = IcdHKLeeImproved().detect(data, sampling_rate_hz=sampling_rate_hz).ic_list_
 
         detected_ics = iterator.results_.ic_list
         snapshot.assert_match(detected_ics, str(tuple(datapoint.group_label)))
