+++ tests/test_initial_contacts/test_icd_shin_improved.py
@@ -4,37 +4,37 @@ import pytest
 from tpcp.testing import TestAlgorithmMixin
 
 from mobgap.data import LabExampleDataset
-from mobgap.icd._hklee_algo_improved import IcdHKLeeImproved
+from mobgap.initial_contacts import IcdShinImproved
 from mobgap.pipeline import GsIterator
 
 
-class TestMetaHKLeeImproved(TestAlgorithmMixin):
+class TestMetaShinImproved(TestAlgorithmMixin):
     __test__ = True
 
-    ALGORITHM_CLASS = IcdHKLeeImproved
+    ALGORITHM_CLASS = IcdShinImproved
 
     @pytest.fixture()
     def after_action_instance(self):
         return self.ALGORITHM_CLASS().detect(
-            pd.DataFrame(np.zeros((1000, 3)), columns=["acc_x", "acc_y", "acc_z"]), sampling_rate_hz=120.0
+            pd.DataFrame(np.zeros((1000, 3)), columns=["acc_x", "acc_y", "acc_z"]), sampling_rate_hz=40.0
         )
 
 
-class TestHKLeeImproved:
+class TestShinImproved:
     def test_invalid_axis_parameter(self):
         with pytest.raises(ValueError):
-            IcdHKLeeImproved(axis="invalid").detect(pd.DataFrame(), sampling_rate_hz=100)
+            IcdShinImproved(axis="invalid").detect(pd.DataFrame(), sampling_rate_hz=100)
 
     def test_no_ics_detected(self):
         data = pd.DataFrame(np.zeros((1000, 3)), columns=["acc_x", "acc_y", "acc_z"])
-        output = IcdHKLeeImproved(axis="x")
-        output.detect(data, sampling_rate_hz=120.0)
+        output = IcdShinImproved(axis="x")
+        output.detect(data, sampling_rate_hz=40.0)
         output_ic = output.ic_list_["ic"]
         empty_output = {}
         assert output_ic.to_dict() == empty_output
 
 
-class TestHKLeeImprovedRegression:
+class TestShinImprovedRegression:
     @pytest.mark.parametrize("datapoint", LabExampleDataset(reference_system="INDIP", reference_para_level="wb"))
     def test_example_lab_data(self, datapoint, snapshot):
         data = datapoint.data_ss
@@ -46,7 +46,7 @@ class TestHKLeeImprovedRegression:
         iterator = GsIterator()
 
         for (gs, data), result in iterator.iterate(data, ref_walk_bouts):
-            result.ic_list = IcdHKLeeImproved().detect(data, sampling_rate_hz=sampling_rate_hz).ic_list_
+            result.ic_list = IcdShinImproved().detect(data, sampling_rate_hz=sampling_rate_hz).ic_list_
 
         detected_ics = iterator.results_.ic_list
         snapshot.assert_match(detected_ics, str(tuple(datapoint.group_label)))
